[ADD] acount_general_ledger_report: Migrated to version 8.0 and create a new Xml for report type Qwed.
