--- TODO-8.0/account_general_ledger_report/i18n/es.po
@@ -1,196 +1,0 @@
-# Translation of OpenERP Server.
-# This file contains the translation of the following modules:
-#	* account_general_ledger_report
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: OpenERP Server 7.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2014-02-18 20:57+0000\n"
-"PO-Revision-Date: 2014-02-18 20:57+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: account_general_ledger_report
-#: view:general.ledger.wiz:0
-msgid "This report allows you to print or generate a pdf of your General Ledger"
-msgstr "Este reporte permite imprimir o generar un pdf del Libro mayor"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,amount_currency:0
-msgid "With Currency"
-msgstr "Con moneda"
-
-#. module: account_general_ledger_report
-#: help:general.ledger.wiz,amount_currency:0
-msgid "It adds the currency column on report if the currency differs from the company currency."
-msgstr "Añade una columna de moneda en el reporte si la moneda es diferente de la moneda de la compañía"
-
-#. module: account_general_ledger_report
-#: code:addons/account_general_ledger_report/wizard/account_general_ledger_report_wizard.py:70
-#, python-format
-msgid "There is no template defined for the selected format. Check if aeroo report exist."
-msgstr "No existe una plantilla definida para el formato seleccionado. Debe existir un reporte de tipo aeroo."
-
-#. module: account_general_ledger_report
-#: view:general.ledger.wiz:0
-msgid "General Ledger"
-msgstr "Libro mayor"
-
-#. module: account_general_ledger_report
-#: help:general.ledger.wiz,historic_strict:0
-msgid "If selected, will display a historical unreconciled lines, taking into account the end of the period or date selected"
-msgstr "Si está selecciondo, desplegará las líneas no conciliadas históricas, dentro de la cuenta al final del período o de la fecha seleccionada"
-
-#. module: account_general_ledger_report
-#: code:addons/account_general_ledger_report/wizard/account_general_ledger_report_wizard.py:70
-#, python-format
-msgid "Error !"
-msgstr "Error !"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,special_period:0
-msgid "Special period"
-msgstr "Período especial"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,sort_selection:0
-msgid "Entries Sorted by"
-msgstr "Entradas ordenadas por"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,company_id:0
-msgid "Company"
-msgstr "Compañía"
-
-#. module: account_general_ledger_report
-#: model:ir.model,name:account_general_ledger_report.model_general_ledger_wiz
-msgid "General Ledger Report Wizard"
-msgstr "Asistente para el Libro Mayor"
-
-#. module: account_general_ledger_report
-#: model:ir.actions.act_window,name:account_general_ledger_report.action_general_ledger_report_wizard
-msgid "General Ledger Report"
-msgstr "Informe del libro mayor"
-
-#. module: account_general_ledger_report
-#: selection:general.ledger.wiz,filter:0
-#: selection:general.ledger.wiz,sort_selection:0
-msgid "Date"
-msgstr "Fecha"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,chart_account_id:0
-msgid "Chart of Account"
-msgstr "Catálogo de Cuentas"
-
-#. module: account_general_ledger_report
-#: view:general.ledger.wiz:0
-#: field:general.ledger.wiz,journal_ids:0
-msgid "Journals"
-msgstr "Diarios"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,target_move:0
-msgid "Target Moves"
-msgstr "Movimientos destino"
-
-#. module: account_general_ledger_report
-#: help:general.ledger.wiz,chart_account_id:0
-msgid "Select Charts of Accounts"
-msgstr "Selecccione el catálogo contable"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,date_to:0
-msgid "End Date"
-msgstr "Fecha final"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,period_from:0
-msgid "Start Period"
-msgstr "Período inicial"
-
-#. module: account_general_ledger_report
-#: selection:general.ledger.wiz,sort_selection:0
-msgid "Name"
-msgstr "Name"
-
-#. module: account_general_ledger_report
-#: selection:general.ledger.wiz,target_move:0
-msgid "All Posted Entries"
-msgstr "Todos los apuntes publicados"
-
-#. module: account_general_ledger_report
-#: help:general.ledger.wiz,fiscalyear_id:0
-msgid "Keep empty for all open fiscal year"
-msgstr "Mantener vacío para todos los años fiscales abiertos"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,period_to:0
-msgid "End Period"
-msgstr "Período final"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,fiscalyear_id:0
-msgid "Fiscal Year"
-msgstr "Año fiscal"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,historic_strict:0
-msgid "Strict History"
-msgstr "Histórico estricto"
-
-#. module: account_general_ledger_report
-#: selection:general.ledger.wiz,filter:0
-msgid "No Filters"
-msgstr "Sin filtros"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,filter:0
-msgid "Filter by"
-msgstr "Filtrar por"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,out_format:0
-msgid "Print Format"
-msgstr "Formato de impresión"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,account_base_report:0
-msgid "Account Base Report"
-msgstr "Informe financiero base"
-
-#. module: account_general_ledger_report
-#: help:general.ledger.wiz,special_period:0
-msgid "Include special period"
-msgstr "Includir períodos especiales"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,account_ids:0
-msgid "Accounts"
-msgstr "Cuentas"
-
-#. module: account_general_ledger_report
-#: selection:general.ledger.wiz,filter:0
-msgid "Periods"
-msgstr "Períodos"
-
-#. module: account_general_ledger_report
-#: field:general.ledger.wiz,date_from:0
-msgid "Start Date"
-msgstr "Fecha inicial"
-
-#. module: account_general_ledger_report
-#: view:general.ledger.wiz:0
-msgid "Account Financial Report"
-msgstr "Informe Financiero"
-
-#. module: account_general_ledger_report
-#: selection:general.ledger.wiz,target_move:0
-msgid "All Entries"
-msgstr "Todos los apuntes"
-
