+++ account_general_ledger_report/report/report.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <data>
+        <!-- QWeb Reports -->
+        <report id="action_report_general_ledger" 
+        model="account.account"
+        string="General Ledger" 
+        report_type="qweb-pdf"
+        name="account_general_ledger_report.report_general_ledger" 
+        file="account_general_ledger_report.report_general_ledger"
+        menu="False"
+        />
+        <record id="action_report_general_ledger" model="ir.actions.report.xml">
+            <field name="paperformat_id" ref="paperformat_us_landscape"/>
+        </record>
+    </data>
+</openerp>
