+++ account_general_ledger_report/wizard/account_general_ledger_report_wizard.xml
@@ -5,7 +5,7 @@
             <field name="name">General Ledger Report </field>
             <field name="model">general.ledger.wiz</field>
             <field name="type">form</field>
-            <field name="inherit_id" ref="account_report_lib.account_report_lib_wizard_view"/>
+            <field name="inherit_id" ref="account_report_lib.account_report_lib_wizard_form_view"/>
             <field name="arch" type="xml">
                 <data>
                     <xpath expr="/form/label[@string='']" position="replace">
